# engines/contracts/engine-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Curf Engine API
  version: 1.0.0
  summary: Data plane for Curf - connections, governed views, queries, policy and audit.
  description: |
    Contract for the Curf engine. Every implementation must pass the shared conformance suite in
    `engines/conformance`. Errors are RFC 9457 `application/problem+json` with a stable `code`.
    Milestone M0 covers identity, audit and operations; later milestones add connections, views,
    queries, reports, exports, sharing and schedules.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Curf engine maintainers
    url: https://github.com/EnterpriseX-Platform/CurfAI
servers:
  - url: http://localhost:8080/engine/v1
    description: Local engine
...
security:
  - bearerAuth: []
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [Schedules]
      summary: Schedules of the workspace
      description: Needs `schedule:manage`.
      responses:
        "200":
          description: Schedules by name.
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Schedule" } }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createSchedule
      tags: [Schedules]
      summary: Send a report on a timetable
      description: |
        Every fire of `cron` (five fields, read in `timezone`, UTC when omitted) runs the report and mails the file to
        `recipients`. It runs **as the person who saved the schedule**, with their row rules and masking, and the same
        rights are checked on every run: they must be able to run the report. Needs `schedule:manage`. Recipients must be
        on `curf.engine.schedules.allowed-recipient-domains`, which is empty by default, so nothing can be mailed out until
        an operator allows a domain. A schedule may not fire more often than every 5 minutes by default.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleRequest" }
      responses:
        "201":
          description: The schedule and its next fire.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Schedule" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/InvalidInput" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthenticated:
      description: Missing, invalid or untrusted token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The caller lacks the required permission.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InvalidInput:
      description: The request is malformed or fails validation. For SQL the code is `CURF_SQL_REJECTED` and `errors` says why.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such resource in the caller's tenant.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string, examples: ["urn:curf:error:forbidden"] }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            - CURF_INVALID_INPUT
            - CURF_SQL_REJECTED
            - CURF_QUERY_TIMEOUT
            - CURF_TOO_MANY_QUERIES
            - CURF_UNAUTHENTICATED
            - CURF_FORBIDDEN
            - CURF_SEGREGATION_OF_DUTIES
            - CURF_NOT_FOUND
            - CURF_STALE_VERSION
            - CURF_CONFLICT
            - CURF_ROW_LIMIT
            - CURF_EXPORT_UNAVAILABLE
            - CURF_CONNECTION_FAILED
            - CURF_INTERNAL
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    ScheduleRequest:
      type: object
      required: [reportId, name, cron, format, recipients]
      properties:
        reportId: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        cron: { type: string, description: "Five fields: minute hour day-of-month month day-of-week.", examples: ["30 6 * * 1-5"] }
        timezone: { type: string, description: IANA zone, default: UTC, examples: [Asia/Bangkok] }
        format: { type: string, enum: [CSV, XLSX, DOCX, PDF] }
        locale: { type: string, enum: [th, en] }
        params: { type: object, description: Report parameter values for every run. }
        recipients: { type: array, maxItems: 20, items: { type: string, format: email } }
        enabled: { type: boolean, default: true }
    Schedule:
      type: object
      properties:
        id: { type: string, format: uuid }
        reportId: { type: string, format: uuid }
        name: { type: string }
        cron: { type: string }
        timezone: { type: string }
        format: { type: string, enum: [CSV, XLSX, DOCX, PDF] }
        locale: { type: [string, "null"] }
        params: { type: object }
        recipients: { type: array, items: { type: string } }
        runAs: { type: string, description: Subject of the person it runs as. }
        enabled: { type: boolean }
        nextRunAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
...